+++ spec/rubocop/cop/style/identical_conditional_branches_spec.rb
@@ -47,6 +47,26 @@ describe RuboCop::Cop::Style::IdenticalConditionalBranches do
     end
   end
 
+  context 'on if..else with identical leading lines' do
+    let(:source) do
+      ['if something',
+       '  do_x',
+       '  method_call_here(1, 2, 3)',
+       'else',
+       '  do_x',
+       '  1 + 2 + 3',
+       'end']
+    end
+
+    it 'registers an offense' do
+      expect(cop.offenses.size).to eq(2)
+      expect(cop.messages).to eq([
+                                   'Move `do_x` out of the conditional.',
+                                   'Move `do_x` out of the conditional.'
+                                 ])
+    end
+  end
+
   context 'on if..elsif with no else' do
     let(:source) do
       ['if something',
@@ -75,7 +95,7 @@ describe RuboCop::Cop::Style::IdenticalConditionalBranches do
     end
   end
 
-  context 'on case with identical trailing lines' do
+  context 'on case with identical bodies' do
     let(:source) do
       ['case something',
        'when :a',
@@ -84,6 +104,56 @@ describe RuboCop::Cop::Style::IdenticalConditionalBranches do
        '  do_x',
        'else',
        '  do_x',
+       'end']
+    end
+
+    it 'registers an offense' do
+      expect(cop.offenses.size).to eq(3)
+      expect(cop.messages).to eq([
+                                   'Move `do_x` out of the conditional.',
+                                   'Move `do_x` out of the conditional.',
+                                   'Move `do_x` out of the conditional.'
+                                 ])
+    end
+  end
+
+  context 'on case with identical trailing lines' do
+    let(:source) do
+      ['case something',
+       'when :a',
+       '  x1',
+       '  do_x',
+       'when :b',
+       '  x2',
+       '  do_x',
+       'else',
+       '  x3',
+       '  do_x',
+       'end']
+    end
+
+    it 'registers an offense' do
+      expect(cop.offenses.size).to eq(3)
+      expect(cop.messages).to eq([
+                                   'Move `do_x` out of the conditional.',
+                                   'Move `do_x` out of the conditional.',
+                                   'Move `do_x` out of the conditional.'
+                                 ])
+    end
+  end
+
+  context 'on case with identical leading lines' do
+    let(:source) do
+      ['case something',
+       'when :a',
+       '  do_x',
+       '  x1',
+       'when :b',
+       '  do_x',
+       '  x2',
+       'else',
+       '  do_x',
+       '  x3',
        'end']
     end